// baselines/scenario-4/frontend/src/generated/booking-client.ts
export interface AuthRequest {
  email: string;
  password: string;
}

export interface AuthUserResponse {
  id: string;
  email: string;
}

export interface CsrfResponse {
  token: string;
}

export interface Table {
  id: string;
  capacity: number;
}

export interface Restaurant {
  id: string;
  name: string;
  cuisine: string;
  neighborhood: string;
  description: string;
  tables: Table[];
}

export interface Booking {
  id: string;
  userId: string;
  restaurantId: string;
  restaurantName: string;
  tableId: string;
  partySize: number;
  date: string;
  time: string;
  guestName: string;
  guestEmail: string;
}

export interface CreateBookingRequest {
  restaurantId: string;
  date: string;
  time: string;
  partySize: number;
  guestName: string;
  guestEmail: string;
}

export interface AvailabilitySlot {
  time: string;
  availableTableCount: number;
}

export interface ErrorResponse {
  code: string;
  message: string;
}